+++ hosts/delta/home.nix
@@ -1,5 +1,0 @@
-{ pkgs, ... }: {
-  home.packages = builtins.attrValues {
-    inherit (pkgs) discord;
-  };
-}